refactor: centralized add question message, remove addMessage action, added Roboto font
