+++ src/App.scss
@@ -1,3 +1,10 @@
+ @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
+ 
+ body {
+     font-family: 'Roboto', sans-serif;
+     /* Apply the font */
+ }
+
 .Container {
     display: flex;
     flex-direction: column;
