--- src/axios.ts.old
@@ -1,18 +1,0 @@
-import axios, { AxiosInstance, InternalAxiosRequestConfig } from 'axios'
-
-const bearerInterceptor = (config: InternalAxiosRequestConfig): InternalAxiosRequestConfig => {
-  if (config.withCredentials === false) {
-    return config
-  }
-
-  config.headers.set('Authorization', 'Bearer MY_KEY}')
-
-  return config
-}
-
-const instance: AxiosInstance = axios.create({
-  baseURL: 'https://api.openai.com/v1/',
-  headers: { 'content-type': 'application/json' },
-})
-
-instance.interceptors.request.use(bearerInterceptor)
